requirements.txt
python-documentcloud==4.0.1
requests
tqdm
patool
bs4
clouddl
